PowerDNS Authoritative Nameserver 4.2.2
I couldn't bump to 4.3.3 because it requires bringing in LUA and several
related LUA dependencies. `tolua`, for example, was a 4.3.3 dependency
that did not work not lend itself to packaging with BOSH.

I also couldn't bump boost to 1.73.0 because it wouldn't compile.

I updated the developer docs to use the vSphere manifest, not the BOSH
Lite manifest, for testing. I always use vSphere, and I think I'm the
only one who reads the developer docs.